refactor package name
package 명 user -> userinfo
