+++ DDUBUCK/app/src/main/AndroidManifest.xml
@@ -23,13 +23,13 @@
 
         </activity>
 
-        <activity android:name=".user.UserInfoBirthdayActivity">
+        <activity android:name=".userinfo.UserInfoBirthdayActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".user.UserInfoHeightWeightActivity"/>
+        <activity android:name=".userinfo.UserInfoHeightWeightActivity"/>
         <activity android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
