+++ DDUBUCK/app/src/main/res/layout/birthday_info_layout.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".user.UserInfoBirthdayActivity">
+    tools:context=".userinfo.UserInfoBirthdayActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/birthday_toolbar"
