+++ DDUBUCK/app/src/main/res/layout/height_weight_info_layout.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".user.UserInfoHeightWeightActivity">
+    tools:context=".userinfo.UserInfoHeightWeightActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:layout_width="match_parent"
